lambdas/yt_upload/requirements.txt
pillar-youtube-upload==0.3.0
pymongo[srv]
s3fs